// project-manager-ui/src/main/web/src/app/shared/model/projmanager.model.ts
export class ProjManagerModel {
    project: string;
    projectId: string;
    task: string;
    enableParentTask: boolean;
    hasStartEndDate: boolean;
    priority: number;
    parentTask: string;
    parentId: string;
    startDate: string;
    endDate: string;
    action: string;
    firstName: string;
    lastName: string;
    employeeId: string;
    userId: string;
    userName: string;
}